Finish day 2

+++ src/template.rs
@@ -1,15 +1,15 @@
 #[aoc_generator(day1)]
-pub fn input_generator(input: &str) -> Vec<u32> {
+pub fn input_generator(input: &str) -> Vec<usize> {
     vec![1]
 }
 
 #[aoc(day1, part1)]
-pub fn solve_part1(input: &[u32]) -> u32 {
+pub fn solve_part1(input: &[u32]) -> usize {
     1
 }
 
 #[aoc(day1, part2)]
-pub fn solve_part2(input: &[u32]) -> u32 {
+pub fn solve_part2(input: &[u32]) -> usize {
     1
 }
